Create programming.md
A first list of skills for the "field" of Programming/Development, lifted from a Twitter discussion by Zed.  
Not sure this fully follows the spirit of the exercise, because Zed in essence asked what skills a "Level 5 Programmer" should have, which would just be the description of lvl 5 for the whole field. Buut oh well, better than nothing.